requirements.txt
black==24.2.0
isort==5.13.2
discord.py==2.3.2
jishaku==2.5.2
python-dotenv==1.0.1
synology_api==0.6.3
gspread==6.0.2
google-api-python-client==2.118.0